Fixed local refrecnes to images and sound files
Put all static file name refrences in start of MainviewModel instead of buried in the methods. Backgroudn image is currently in the constructor as thats where it needs to be for the time being. Not ideal.

Refercnes to Directory have been simplieifed and nopw only point to one call made int he now FileRepositoryClass (was TextFile Repository)

Some other tidying up- removed unused files, renamed a couple of things.